fixes + path change
