+++ cupy/__init__.py
@@ -306,7 +306,6 @@ from cupy.manipulation.join import dstack  # NOQA
 from cupy.manipulation.join import hstack  # NOQA
 from cupy.manipulation.join import stack  # NOQA
 from cupy.manipulation.join import vstack  # NOQA
-from cupy.manipulation.join import append  # NOQA
 
 from cupy.manipulation.kind import asfortranarray  # NOQA
 from cupy.manipulation.kind import require  # NOQA
@@ -321,6 +320,7 @@ from cupy.manipulation.tiling import repeat  # NOQA
 from cupy.manipulation.tiling import tile  # NOQA
 
 from cupy.manipulation.add_remove import unique  # NOQA
+from cupy.manipulation.add_remove import append  # NOQA
 
 from cupy.manipulation.rearrange import flip  # NOQA
 from cupy.manipulation.rearrange import fliplr  # NOQA
