+++ cupy/manipulation/add_remove.py
@@ -1,5 +1,7 @@
 import cupy
-
+from cupy.manipulation.shape import ravel
+from cupy.creation.from_data import asanyarray
+from cupy import core
 
 # TODO(okuta): Implement delete
 
@@ -7,7 +9,33 @@ import cupy
 # TODO(okuta): Implement insert
 
 
-# TODO(okuta): Implement append
+def append(tup, values, axis=None):
+    """
+       Append values to the end of an array.
+
+       Args:
+         tup (sequence of arrays): Arrays to which values are appended
+         values :
+          array of values that are going to be appended.
+             It must be of the correct shape (the same shape as `tup`).
+             If `axis` is not specified, `values` are flattened before use.
+         axis : int, optional
+           The axis along which `values` are appended.  If `axis` is not
+           given, both `tup` and `values` are flattened before use.
+
+       Returns:
+          cupy.ndarray : `tup` with `values` appended to `axis`.
+
+       .. seealso:: :func:`numpy.append`
+    """
+    arr = asanyarray(tup)
+    values = asanyarray(values)
+    if axis is None:
+        if arr.ndim != 1:
+            arr = arr.ravel()
+        values = ravel(values)
+        return core.concatenate_method((arr, values), 0)
+    return core.concatenate_method((arr, values), axis=axis)
 
 
 # TODO(okuta): Implement resize
